Extract a class for lookup of links
By extracting the class we can add functionality and tests more easily.
It's also easier to understand with some documentation.
